+++ JavaRushTasks/1.JavaSyntax/src/com/javarush/task/task10/task1003/Solution.java
@@ -7,8 +7,9 @@ package com.javarush.task.task10.task1003;
 public class Solution {
     public static void main(String[] args) {
         float f = (float) 128.50;
-        int i = (int) f;
+        int i = (byte) f; // was (int)
         int b = (int) (i + f);
+
         System.out.println(b);
     }
 }
