+++ JavaRushTasks/1.JavaSyntax/src/com/javarush/task/task10/task1005/Solution.java
@@ -7,8 +7,11 @@ package com.javarush.task.task10.task1005;
 public class Solution {
     public static void main(String[] args) {
         int a = (byte) 44;
-        int b = (byte) 300;
-        short c = (byte) (b - a);
+        int b = (short) 300;
+        short c = (short) (b - a);
+
+        //System.out.println(a);
+        //System.out.println(b);
         System.out.println(c);
     }
 }
