+++ JavaRushTasks/1.JavaSyntax/src/com/javarush/task/task10/task1008/Solution.java
@@ -8,7 +8,7 @@ public class Solution {
     public static void main(String[] args) {
         int a = 50;
         int b = 17;
-        double d = a / b;
+        double d = (double) a / b;
         System.out.println(d);
     }
 }
